# application-api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Java Starter Kit API
  description: API for Java Starter Kit with Angular frontend
  version: 1.0.0
  contact:
    name: Starter Team
    email: contact@rawz06.fr

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://localhost:4200
    description: Angular dev server (proxied)

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Public
    description: Public endpoints (no authentication required)
  - name: User
    description: User profile endpoints
  - name: Admin
    description: Admin endpoints (requires ADMIN role)
  - name: Batch
    description: Batch job monitoring endpoints

security:
  - BearerAuth: []

paths:
  # ============== PUBLIC ENDPOINTS ==============
  /api/public/info:
    get:
      tags:
        - Public
      summary: Get public application information
      operationId: getPublicInfo
      security: []
      responses:
        '200':
          description: Application information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInfoDto'

  /api/public/auth/login:
    post:
      tags:
        - Auth
      summary: Login to the application
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  # ============== AUTH ENDPOINTS ==============
  /api/auth/me:
    get:
      tags:
        - Auth
      summary: Get current authenticated user
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  # ============== USER PROFILE ENDPOINTS ==============
  /api/user/profile:
    get:
      tags:
        - User
      summary: Get user profile
      operationId: getUserProfile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

    put:
      tags:
        - User
      summary: Update user profile
      operationId: updateUserProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequestDto'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  # ============== ADMIN DASHBOARD ENDPOINTS ==============
  /api/admin/dashboard:
    get:
      tags:
        - AdminDashboard
      summary: Get admin dashboard data
      operationId: getAdminDashboard
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardDto'

  /api/admin/profile:
    get:
      tags:
        - AdminDashboard
      summary: Get admin profile (legacy endpoint)
      operationId: getAdminProfile
      responses:
        '200':
          description: Admin profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProfileDto'

  # ============== ADMIN USER MANAGEMENT ENDPOINTS ==============
  /api/admin/users:
    get:
      tags:
        - AdminUsers
      summary: Get all users
      operationId: getAllUsers
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'

    post:
      tags:
        - AdminUsers
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequestDto'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  /api/admin/users/{id}:
    get:
      tags:
        - AdminUsers
      summary: Get user by ID
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: User not found

    put:
      tags:
        - AdminUsers
      summary: Update user
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequestDto'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: User not found

    delete:
      tags:
        - AdminUsers
      summary: Delete user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageDto'
        '400':
          description: Cannot delete user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  # ============== BATCH JOB ENDPOINTS ==============
  /api/admin/batch/jobs:
    get:
      tags:
        - Batch
      summary: Get all batch jobs
      operationId: getAllBatchJobs
      responses:
        '200':
          description: List of batch jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BatchJobDto'

  /api/admin/batch/jobs/{jobName}/executions:
    get:
      tags:
        - Batch
      summary: Get job executions
      operationId: getJobExecutions
      parameters:
        - name: jobName
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: List of job executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobExecutionDto'

  /api/admin/batch/jobs/{jobName}/executions/{executionId}:
    get:
      tags:
        - Batch
      summary: Get job execution details
      operationId: getJobExecutionDetails
      parameters:
        - name: jobName
          in: path
          required: true
          schema:
            type: string
        - name: executionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Job execution details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobExecutionDetailsDto'
        '404':
          description: Execution not found

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # ============== COMMON DTOs ==============
    ErrorDto:
      type: object
      properties:
        error:
          type: string
          description: Error message

    MessageDto:
      type: object
      properties:
        message:
          type: string
          description: Success message

    # ============== AUTH DTOs ==============
    LoginRequestDto:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1

    LoginResponseDto:
      type: object
      properties:
        token:
          type: string
          description: JWT token

    # ============== PUBLIC DTOs ==============
    AppInfoDto:
      type: object
      properties:
        appName:
          type: string
        version:
          type: string
        description:
          type: string

    # ============== USER DTOs ==============
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time

    UserProfileDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time

    UpdateProfileRequestDto:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        currentPassword:
          type: string
        newPassword:
          type: string

    CreateUserRequestDto:
      type: object
      required:
        - login
        - password
        - email
      properties:
        login:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
          default: USER
        active:
          type: boolean
          default: true

    UpdateUserRequestDto:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        active:
          type: boolean

    # ============== ADMIN DTOs ==============
    DashboardDto:
      type: object
      properties:
        message:
          type: string
        user:
          type: string
        stats:
          $ref: '#/components/schemas/DashboardStatsDto'

    DashboardStatsDto:
      type: object
      properties:
        totalUsers:
          type: integer
          format: int64
        activeUsers:
          type: integer
          format: int64
        adminUsers:
          type: integer
          format: int64
        recentUsers:
          type: integer
          format: int64

    AdminProfileDto:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        role:
          type: string

    # ============== BATCH DTOs ==============
    BatchJobDto:
      type: object
      properties:
        name:
          type: string
        instanceCount:
          type: integer
          format: int64
        lastExecutionStatus:
          type: string
        lastExecutionTime:
          type: string
          format: date-time
        lastExecutionEndTime:
          type: string
          format: date-time

    JobExecutionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        instanceId:
          type: integer
          format: int64
        jobName:
          type: string
        status:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        exitCode:
          type: string
        exitDescription:
          type: string

    JobExecutionDetailsDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        jobName:
          type: string
        status:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        exitCode:
          type: string
        exitDescription:
          type: string
        createTime:
          type: string
          format: date-time
        lastUpdated:
          type: string
          format: date-time
        steps:
          type: array
          items:
            $ref: '#/components/schemas/StepExecutionDto'

    StepExecutionDto:
      type: object
      properties:
        stepName:
          type: string
        status:
          type: string
        readCount:
          type: integer
          format: int32
        writeCount:
          type: integer
          format: int32
        commitCount:
          type: integer
          format: int32
        rollbackCount:
          type: integer
          format: int32
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time